Change `operator&` in `FlagEnumClassOperations` to return the flags
This operator was defined to return a boolean because it is what is
actually needed in most situations. However, this behavior is very
unintuitive as well because `&` is supposed to be a binary and. So this was
changed and can be used by defining
`CPP_UTILITIES_FLAG_ENUM_CLASS_NO_LEGACY_AND`. Then `&&` needs to be used
if a boolean is needed.